// src/enums.ts
export enum GPTModel {
    "GPT 4o" = "gpt-4o",
    "GPT 3.5 Turbo" = "gpt-3.5-turbo",
}

export enum DialogRole {
    System = "system",
    Bot = "assistant",
    User = "user",
}

export enum KeyboardActions {
    SelectModel = "select-model",
    EnableStreaming = "enable-streaming",
    DisableStreaming = "disable-streaming",
}